Updated Time Tracking
[Change] afc_sound now tracks the time it takes for the web clients to recieve the audio file in a play-ready state in addition to the file playback length.

Experiment CFG files can now additionally define a new def field (def.trialResponseIntervalTime) to add a fixed number of seconds to the measured time for each trial to allow for participant response.

This time is added every trial regardless of whether the participant took more or less time to respond (actual response time is not tracked).